+++ uui/src/lib/shared/theme/config/theme-config.types.d.ts
@@ -1,3 +1,16 @@
+import {
+  Properties,
+  PropertiesHyphen,
+  type ObsoleteProperties,
+  type StandardProperties,
+  type SvgProperties
+} from 'csstype';
+
+export type CssProperties = StandardProperties &
+  SvgProperties &
+  ObsoleteProperties &
+  PropertiesHyphen;
+
 export interface CssObject {
-  [key: string]: Record<string, string>;
+  [key: string]: CssProperties;
 }