+++ uui/src/lib/shared/theme/typography-config/typography-config.utils.ts
@@ -1,10 +1,7 @@
-import type { TypographyVariant } from '../../../Typography/typography.types';
 import { camelBackToDash } from '../config/theme-config.utils';
-import type {
-  ThemeTypography,
-  TypographyConfig,
-  TypographyPaletteValues
-} from './typography-config.types';
+import type { CssProperties } from '../config/theme-config.types';
+import type { TypographyVariant } from '../../../Typography/typography.types';
+import type { ThemeTypography, TypographyConfig } from './typography-config.types';
 
 const TYPOGRAPHY_PREFIX = 'mdc-typography';
 const TYPOGRAPHY_VARS_PREFIX = `--${TYPOGRAPHY_PREFIX}`;
@@ -12,7 +9,7 @@ const TYPOGRAPHY_VARS_PREFIX = `--${TYPOGRAPHY_PREFIX}`;
 export function createTypographyConfig(typography: ThemeTypography): TypographyConfig {
   const typographyConfig: TypographyConfig = { vars: {}, classes: {} };
 
-  Object.entries(typography).forEach(([key, val]: [string, TypographyPaletteValues]) => {
+  Object.entries(typography).forEach(([key, val]: [string, CssProperties]) => {
     const entryConfig = createTypographyClasses(val, key);
     typographyConfig.classes = { ...typographyConfig.classes, ...entryConfig.classes };
     typographyConfig.vars = { ...typographyConfig.vars, ...entryConfig.vars };
@@ -22,7 +19,7 @@ export function createTypographyConfig(typography: ThemeTypography): TypographyC
 }
 
 export function createTypographyClasses(
-  typographyPalette: TypographyPaletteValues,
+  typographyPalette: CssProperties,
   prefix: string
 ): TypographyConfig {
   const typoConfig: TypographyConfig = { classes: {}, vars: {} };
@@ -30,7 +27,7 @@ export function createTypographyClasses(
   const dashedPrefix = camelBackToDash(prefix);
   const typographyClass = createTypographyConfigClass(dashedPrefix);
 
-  Object.entries(typographyPalette).forEach(([key, val]: [string, string]) => {
+  Object.entries(typographyPalette).forEach(([key, val]: [string, unknown]) => {
     const dashedKey = camelBackToDash(key);
     const unPrefixedVarName = `${dashedPrefix}-${dashedKey}`;
     const uuiVarName = createUUITypographyVarName(unPrefixedVarName);
@@ -38,7 +35,7 @@ export function createTypographyClasses(
     const mdcVarValue = createMDCVarValue(uuiVarName);
 
     classValues[dashedKey] = `var(${uuiVarName}, ${val})`;
-    typoConfig.vars[uuiVarName] = val;
+    typoConfig.vars[uuiVarName] = `${val}`;
     typoConfig.vars[mdcVarName] = mdcVarValue;
   });