+++ uui/tests/lib/shared/theme/config/theme-config.utils.test.ts
@@ -5,7 +5,7 @@ import {
 } from '../../../../../src/lib/shared/theme/config/theme-config.utils';
 describe('makeStyles', () => {
   it('creates a css class string from cssObject', () => {
-    expect(makeStyles({ '.mdc-typography--custom-typography': { 'font-size': '1.2px' } })).toBe(
+    expect(makeStyles({ '.mdc-typography--custom-typography': { fontSize: '1.2px' } })).toBe(
       '.mdc-typography--custom-typography{font-size:1.2px;}'
     );
   });